refactor(LottoGenerator): 이름 변경 및 외부의 public 상수 사용
